feat: keep controller between avatar reloads
